+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The `--border` flag can be used to create a border around the ascii image
+- The `--flipX` flag can be used horizontally flip the image
+- The `--flipY` flag can be used vertically flip the image
 - Two more tests, which fully compare the results
 
 ## [0.5.1] - 2022-03-14